refactor: extract transformation/read-input-source-map-file
