+++ packages/babel-core/src/transformation/read-input-source-map-file-browser.ts
@@ -1,0 +1,5 @@
+export default function readInputSourceMapFile(): never {
+  throw new Error(
+    "Reading input source map files is not supported in browsers",
+  );
+}
